requirements.txt
# OpenAI Realtime Audio Testing Dependencies
websockets>=11.0.0
python-dotenv>=1.0.0
# Optional for additional functionality
requests>=2.31.0
# Official OpenAI client with realtime extras
openai[realtime]>=1.14.0
# Web interface dependencies
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
# Audio conversion (same as working simple_browser_test.py)
ffmpeg-python>=0.2.0
numpy
duckduckgo-search 
requests 
beautifulsoup4
playwright==1.*
# ddgs is the renamed duckduckgo_search helper
ddgs
httpx>=0.25.0

# After installing packages, make sure browser binaries are present:
#   python -m playwright install chromium